added tests

+++ Kingdom/apps/character/views.py
@@ -1,4 +1,3 @@
-from django.db.models import Prefetch
 from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
@@ -10,8 +9,7 @@ from apps.character.serializers import CharacterOverallSerializer, CharacterDeta
     AddItemSerializer, EquipItemSerializer, SecondaryStatsSerializer, LevelUpSerializer, SetStatsSerializer, \
     SetStatsDisplaySerializer, SetSpeedSerializer, SetMasterySerializer, CharacterSkillSerializer, \
     CharacterSkillMasterySerializer, SetFeatSerializer, SetSpellSerializer, SetConditionSerializer, \
-    UnEquipSerializer, UnEquipWornItemSerializer
-
+    UnEquipSerializer
 from apps.character.models import Character, SecondaryStats, CharacterStats, CharacterSkillList, CharacterSkillMastery,\
     CharacterFeatList, SpellList, DefenceAndVulnerabilityDamage, CharacterBag, InventoryItems
 from apps.permissions import IsOwner
@@ -45,9 +43,11 @@ class CharacterCreateView(APIView):
     serializer_class = CharacterSerializer
 
     def post(self, request):
-        serializer = self.serializer_class(data=request.data)
+        serializer = self.serializer_class(data=request.data,
+                                           context={'user': request.user}
+                                           )
         serializer.is_valid(raise_exception=True)
-        serializer.save(user=self.request.user)
+        serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
@@ -62,17 +62,20 @@ class SetStatsView(APIView):
         serializer = self.serializer_class_display(character)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def put(self, request, character_id):
-         character = get_object_or_404(CharacterStats.objects.select_related('character'), character_id=character_id)
-         serializer = self.serializer_class(character, data=request.data, partial=True)
-         serializer.is_valid(raise_exception=True)
-         serializer.save()
-         return Response(serializer.data, status=status.HTTP_200_OK)
+    def patch(self, request, character_id):
+        character = get_object_or_404(CharacterStats.objects.select_related('character'), character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, character):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
+        serializer = self.serializer_class(character, data=request.data, partial=True)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 class SetSpeedView(APIView):
     """ Set character's speed """
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = SetSpeedSerializer
 
     def get(self, request, character_id):
@@ -82,6 +85,9 @@ class SetSpeedView(APIView):
 
     def patch(self, request, character_id):
         character = get_object_or_404(CharacterStats, character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, character):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(character, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -90,7 +96,7 @@ class SetSpeedView(APIView):
 
 class SetMasteryView(APIView):
     """ Change mastery levels """
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = SetMasterySerializer
 
     def get(self, request, character_id):
@@ -98,8 +104,11 @@ class SetMasteryView(APIView):
         serializer = self.serializer_class(character)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def put(self, request, character_id):
+    def patch(self, request, character_id):
         character = get_object_or_404(CharacterStats, character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, character):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(character, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -108,7 +117,7 @@ class SetMasteryView(APIView):
 
 class AddItemView(APIView):
     """ Add new item in inventory """
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = AddItemSerializer
 
     def get(self, request, character_id):
@@ -120,6 +129,9 @@ class AddItemView(APIView):
     def post(self, request, character_id):
         bag = get_object_or_404(CharacterBag.objects.select_related('character').
                                 prefetch_related('inventory'), character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, bag):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(data=request.data,
                                            context={"inventory": bag})
         serializer.is_valid(raise_exception=True)
@@ -134,6 +146,9 @@ class RemoveItemView(APIView):
 
     def delete(self, request, item_id):
         item = get_object_or_404(InventoryItems.objects.select_related('bag'), id=item_id)
+        if not IsOwner().has_object_permission(request, self, item.bag):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         item.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -144,10 +159,14 @@ class EquipItemView(APIView):
     serializer_class = EquipItemSerializer
 
     def post(self, request, *args, **kwargs):
+        bag = get_object_or_404(CharacterBag, id=request.data.get('bag_id'))
+        if not IsOwner().has_object_permission(request, self, bag):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
-        return Response({"message": f"item equipped"}, status=status.HTTP_200_OK)
+        return Response({"message": f"item equipped"}, status=status.HTTP_201_CREATED)
 
 
 class BaseUnEquipView(APIView):
@@ -157,6 +176,9 @@ class BaseUnEquipView(APIView):
 
     def patch(self, request, character_id):
         equipped_items = get_object_or_404(EquippedItems, bag_id=character_id)
+        if not IsOwner().has_object_permission(request, self, equipped_items):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(data=request.data)
         serializer.is_valid(raise_exception=True)
         if serializer.validated_data.get('un_equip'):
@@ -196,7 +218,7 @@ class UnEquipSecondWeaponView(BaseUnEquipView):
 class SetSecondaryStatsView(APIView):
     """ Change characteristics (health and e.t.c) """
 
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = SecondaryStatsSerializer
 
     def get(self, request, character_id):
@@ -206,6 +228,9 @@ class SetSecondaryStatsView(APIView):
 
     def patch(self, request, character_id):
         secondary_stats = get_object_or_404(SecondaryStats, character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, secondary_stats):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(secondary_stats, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -217,6 +242,9 @@ class DeleteWornItemView(APIView):
 
     def delete(self, request, character_id, item_id):
         equipped_items = get_object_or_404(EquippedItems, bag_id=character_id)
+        if not IsOwner().has_object_permission(request, self, equipped_items):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         item_to_remove = get_object_or_404(equipped_items.worn_items, pk=item_id)
         equipped_items.worn_items.remove(item_to_remove)
         return Response({"message": "Item removed from worn items successfully"}, status=status.HTTP_204_NO_CONTENT)
@@ -224,11 +252,14 @@ class DeleteWornItemView(APIView):
 
 class SecondaryStatsView(APIView):
     """ Display secondary characteristics """
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = SecondaryStatsSerializer
 
     def get(self, request, character_id):
         secondary_stats = get_object_or_404(SecondaryStats, character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, secondary_stats):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(secondary_stats)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
@@ -236,7 +267,7 @@ class SecondaryStatsView(APIView):
 class LevelUpView(APIView):
     """ Change character's level """
 
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = LevelUpSerializer
 
     def get(self, request, character_id):
@@ -247,6 +278,9 @@ class LevelUpView(APIView):
     def patch(self, request, character_id):
         character = get_object_or_404(Character.objects.select_related
                                       ('class_player', 'character_stats', 'secondary_stats'), pk=character_id)
+        if not IsOwner().has_object_permission(request, self, character):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(character, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -256,11 +290,14 @@ class LevelUpView(APIView):
 class SetSkillsView(APIView):
     """ Set character's skills """
 
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = CharacterSkillSerializer
 
-    def put(self, request, character_id):
+    def patch(self, request, character_id):
         character_skills = get_object_or_404(CharacterSkillList, character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, character_skills):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(character_skills, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -270,7 +307,7 @@ class SetSkillsView(APIView):
 class SetSkillMasteryView(APIView):
     """ Set Character's skill mastery """
 
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = CharacterSkillMasterySerializer
 
     def get(self, request, skill_id):
@@ -280,6 +317,9 @@ class SetSkillMasteryView(APIView):
 
     def patch(self, request, skill_id):
         mastery = get_object_or_404(CharacterSkillMastery.objects.select_related('skill_list'), id=skill_id)
+        if not IsOwner().has_object_permission(request, self, mastery):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(mastery, data=request.data, partial=True)
         serializer.is_valid()
         serializer.save()
@@ -289,15 +329,13 @@ class SetSkillMasteryView(APIView):
 class SetFeatView(APIView):
     """ Set character's feat list """
 
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = SetFeatSerializer
 
     def get(self, request, character_id):
         feat_list = get_object_or_404(CharacterFeatList.objects.
                                       select_related('character').
                                       prefetch_related('feat_class'), character_id=character_id)
-        print(feat_list.character.class_player)
-        print(feat_list.character.level)
         serializer = self.serializer_class(feat_list)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
@@ -305,6 +343,9 @@ class SetFeatView(APIView):
         feat_list = get_object_or_404(CharacterFeatList.objects.
                                       select_related('character').
                                       prefetch_related('feat_class'), character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, feat_list):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(feat_list,
                                            data=request.data,
                                            context={"class_player":feat_list.character.class_player,
@@ -315,7 +356,7 @@ class SetFeatView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
 
-class SetSpellView(APIView):
+class SetSpellView(APIView, IsOwner):
     """ Set character's spell list """
 
     permission_classes = [IsAuthenticated]
@@ -328,6 +369,9 @@ class SetSpellView(APIView):
 
     def patch(self, request, character_id):
         feat_list = get_object_or_404(SpellList, character_id=character_id)
+        if not IsOwner().has_object_permission(request, self, feat_list):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializer_class(feat_list, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -337,7 +381,7 @@ class SetSpellView(APIView):
 class SetCondition(APIView):
     """ Set character's defence and vulnerability """
 
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsOwner]
     serializer_class = SetConditionSerializer
 
     def get(self, request, character_id):
@@ -349,8 +393,10 @@ class SetCondition(APIView):
     def patch(self, request, character_id):
         conditions = get_object_or_404(DefenceAndVulnerabilityDamage.objects.
                                        select_related('character'), character_id=character_id)
-        serializer = self.serializer_class(conditions, data=request.data, partial=True)
+        if not IsOwner().has_object_permission(request, self, conditions):
+            return Response({"error": "You do not have permission to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
+        serializer = self.serializer_class(conditions, data=request.data)
         serializer.is_valid(raise_exception=True)
-        serializer.validated_data['character_id'] = character_id
         serializer.save()
         return Response(serializer.data, status=status.HTTP_200_OK)
